Wrap up the login box into a component
